FA-39 toogle sidebar at click
