+++ frontend/src/pages/layout/LayoutNew.tsx
@@ -3,45 +3,24 @@
 import React, { useState } from "react";
 import { Outlet, useLocation } from "react-router-dom";
 import styles from "./_Layout.module.css";
-import { useAppContext } from "../../providers/AppProviders";
 import Sidebar from "../../components/Sidebar/Sidebar";
 import Navbar from "../../components/Navbar/NavBar";
 
 const Layout: React.FC = () => {
-    const { showHistoryPanel, setShowHistoryPanel, showFeedbackRatingPanel, setShowFeedbackRatingPanel, settingsPanel, setSettingsPanel } = useAppContext();
-
     const { pathname } = useLocation();
-    const [isCollapsed, setIsCollapsed] = useState(false);
-
-    const handleShowHistoryPanel = () => {
-        setShowHistoryPanel(!showHistoryPanel);
-        setShowFeedbackRatingPanel(false);
-        setSettingsPanel(false);
-    };
-
-    const handleShowFeedbackRatingPanel = () => {
-        setShowFeedbackRatingPanel(!showFeedbackRatingPanel);
-        setSettingsPanel(false);
-        setShowHistoryPanel(false);
-    };
-
-    const handleShowSettings = () => {
-        setSettingsPanel(!settingsPanel);
-        setShowHistoryPanel(false);
-        setShowFeedbackRatingPanel(false);
-    };
+    const [isCollapsed, setIsCollapsed] = useState(true);
 
     return (
         <>
             {/* Sidebar Start */}
-            <Sidebar />
+            <Sidebar isCollapsed={isCollapsed} setIsCollapsed={setIsCollapsed} />
             {/* Sidebar End */}
 
             {/* Main Wrapper */}
             <div className={`${styles.bodyWrapper} ${styles.bodyWrapperFull}`}>
                 {/* Header Start */}
                 <header className={`${styles.appHeader} ${styles.appHeaderFull} border-bottom py-2 bg-light`}>
-                    <Navbar />
+                    <Navbar setIsCollapsed={setIsCollapsed} />
                 </header>
                 {/* Header End */}
